+++ src/client/sprites/Button.ts
@@ -1,4 +1,5 @@
 import { canvas, ctx, penCtx } from '@main/canvas.ts';
+import Engine from '@main/Engine.ts';
 import Sprite, { type BoundingBox, type SpriteOptions } from '@main/Sprite.ts';
 import TSCMath from '@main/TSCMath.ts';
 
@@ -54,8 +55,8 @@ export default class Button extends Sprite {
         const height = 2 * (Math.abs(w * sin) + Math.abs(h * cos)) * this.size;
 
         return {
-            x: this.x,
-            y: this.y,
+            x: this.x - Engine.camera.x,
+            y: this.y + Engine.camera.y,
             width, height
         };
     }
@@ -78,8 +79,8 @@ export default class Button extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
+        const cX = this.x + canvas.width / 2 - Engine.camera.x;
+        const cY = -this.y + canvas.height / 2 + Engine.camera.y;
         c.translate(cX, cY);
         c.rotate(this.toRadians(this.dir));
         c.translate(-this.pivot[0], this.pivot[1]);