+++ src/client/sprites/Circle.ts
@@ -1,3 +1,4 @@
+import Engine from '@main/Engine.ts';
 import Sprite, { type BoundingBox, type SpriteOptions } from '@main/Sprite.ts';
 import { canvas, ctx, penCtx } from '@main/canvas.ts';
 
@@ -20,8 +21,8 @@ export default class Circle extends Sprite {
 
     public getBoundingBox(): BoundingBox {
         return {
-            x: this.x,
-            y: this.y,
+            x: this.x - Engine.camera.x,
+            y: this.y + Engine.camera.y,
             width: this.radius * 2 * this.size,
             height: this.radius * 2 * this.size
         };
@@ -46,8 +47,8 @@ export default class Circle extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
+        const cX = this.x + canvas.width / 2 - Engine.camera.x;
+        const cY = -this.y + canvas.height / 2 + Engine.camera.y;
         c.translate(cX, cY);
         c.rotate(this.toRadians(this.dir));
         c.translate(-this.pivot[0], this.pivot[1]);