+++ src/client/sprites/CustomPolygon.ts
@@ -2,6 +2,7 @@ import Sprite, { type BoundingBox, type SpriteOptions } from '@main/Sprite.ts';
 import { canvas, ctx, penCtx } from '@main/canvas.ts';
 import type { Vec2 } from '@ctypes/Vectors.ts';
 import TSCMath from '@main/TSCMath.ts';
+import Engine from '@main/Engine.ts';
 
 export interface CustomPolygonOptions extends SpriteOptions {
     vertices?: Vec2[];
@@ -61,8 +62,8 @@ export default class CustomPolygon extends Sprite {
         const y = minY + height / 2;
 
         return {
-            x: x * this.size,
-            y: y * this.size,
+            x: x * this.size - Engine.camera.x,
+            y: y * this.size + Engine.camera.y,
             width: width * this.size,
             height: height * this.size
         };
@@ -89,12 +90,12 @@ export default class CustomPolygon extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
+        const cX = this.x + canvas.width / 2 - Engine.camera.x;
+        const cY = -this.y + canvas.height / 2 + Engine.camera.y;
         c.translate(cX, cY);
         c.rotate(this.toRadians(this.dir));
         c.translate(-this.pivot[0], this.pivot[1]);
-TSCMath
+
         const path = this.getCachedPath();
 
         c.fillStyle = this.color;