+++ src/client/sprites/RegularPolygon.ts
@@ -1,4 +1,5 @@
 import { canvas, ctx, penCtx } from '@main/canvas.ts';
+import Engine from '@main/Engine.ts';
 import Sprite, { type BoundingBox, type SpriteOptions } from '@main/Sprite.ts';
 import TSCMath from '@main/TSCMath.ts';
 
@@ -23,8 +24,8 @@ export default class RegularPolygon extends Sprite {
 
     public getBoundingBox(): BoundingBox {
         return {
-            x: this.x,
-            y: this.y,
+            x: this.x - Engine.camera.x,
+            y: this.y + Engine.camera.y,
             width: this.radius * 2 * this.size,
             height: this.radius * 2 * this.size
         };
@@ -51,8 +52,8 @@ export default class RegularPolygon extends Sprite {
 
         c.save();
 
-        const cX = this.x + canvas.width / 2;
-        const cY = -this.y + canvas.height / 2;
+        const cX = this.x + canvas.width / 2 - Engine.camera.x;
+        const cY = -this.y + canvas.height / 2 + Engine.camera.y;
         c.translate(cX, cY);
         c.rotate(TSCMath.toRadians(this.dir));
         c.translate(-this.pivot[0], this.pivot[1]);